Revert "Revert "Move shared models to core model module""
This reverts commit 280f5955
