+++ core/model/src/commonMain/kotlin/com/github/kr328/clash/core/model/AccessControlMode.kt
@@ -1,4 +1,4 @@
-package com.github.kr328.clash.service.model
+package com.github.kr328.clash.core.model
 
 enum class AccessControlMode {
   AcceptAll,
