+++ core/model/src/commonMain/kotlin/com/github/kr328/clash/core/model/DarkMode.kt
@@ -1,4 +1,4 @@
-package com.github.kr328.clash.common.model
+package com.github.kr328.clash.core.model
 
 enum class DarkMode {
   Auto,
