+++ glue/build.gradle.kts
@@ -4,6 +4,7 @@ plugins {
 
 dependencies {
   api(projects.core)
+  api(projects.core.model)
   api(projects.service)
   api(projects.common)
