+++ settings.gradle.kts
@@ -51,6 +51,7 @@ enableFeaturePreview("STABLE_CONFIGURATION_CACHE")
 include(
   ":androidApp",
   ":core",
+  ":core:model",
   ":service",
   ":common",
   ":glue",
