+++ ui/build.gradle.kts
@@ -5,6 +5,7 @@ plugins {
 kotlin {
   sourceSets {
     commonMain.dependencies {
+      implementation(projects.core.model)
       implementation(projects.common)
     }
   }
